rag-pipeline/requirements.txt
# Python 3.10
beautifulsoup4==4.13.3
requests==2.32.3
transformers==4.48.3
sentence-transformers==3.4.1
accelerate==1.3.0
bitsandbytes==0.42.0
ipykernel==6.29.5
langchain==0.3.18
faiss-cpu==1.10.0
langchain-community==0.3.17
langchain-huggingface==0.1.2
torch==2.1.1 --index-url https://download.pytorch.org/whl/torch_stable.html
pypdf==5.3.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
cryptography==44.0.1
prometheus_client==0.21.1
pytest==8.3.5
pytest-mock==3.14.0
pytest-cov==6.0.0
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-exporter-jaeger==1.21.0